list sorted based on updateDate

+++ GetirTodo/Application/ToDoItemDetail/ToDoItemDetailVM.swift
@@ -9,10 +9,14 @@ import Foundation
 
 final class ToDoItemDetailVM {
     private let toDoListItem: ToDoListItem
+    private let initialTitle: String
+    private let inititalDetail: String
     
     init(item: ToDoListItem?) {
         if let item = item {
             self.toDoListItem = item
+            self.initialTitle = item.title
+            self.inititalDetail = item.detail
         } else {
             self.toDoListItem = ToDoListItem()
             RealmManager.shared.addObject(object: toDoListItem) {
@@ -20,7 +24,8 @@ final class ToDoItemDetailVM {
             } errorHandler: { error in
                 print(error)
             }
-
+            self.initialTitle = ""
+            self.inititalDetail = ""
         }
     }
     
@@ -32,12 +37,18 @@ final class ToDoItemDetailVM {
         toDoListItem.detail
     }
     
-    func updateItemIfNotEmpty(title: String, detail: String) {
-        if isItemEmpty(title, detail) {
-            deleteItem()
-        } else {
-            updateItem(title, detail)
+    func checkAndUpdateItem(title: String, detail: String) {
+        if IsItemChanged(title, detail) {
+            if isItemEmpty(title, detail) {
+                deleteItem()
+            } else {
+                updateItem(title, detail)
+            }
         }
+    }
+    
+    private func IsItemChanged(_ title: String, _ detail: String) -> Bool {
+        return initialTitle != title || inititalDetail != detail
     }
 
     private func isItemEmpty(_ title: String, _ detail: String) -> Bool {
@@ -56,7 +67,6 @@ final class ToDoItemDetailVM {
         RealmManager.shared.updateObject {
             toDoListItem.title = title
             toDoListItem.detail = detail
-
         } successHandler: {
             print("UPDATED")
         } errorHandler: { error in
